# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Restaurante",
        "description": "API para la gestión de restaurantes",
        "contact": {
            "name": "Soporte API",
            "email": "soporte@restaurante.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/reservas": {
            "get": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Lista todas las reservas",
                "operationId": "d31270690bd9559009b7c90e03ee0fe0",
                "parameters": [
                    {
                        "name": "fecha",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "estado",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "mesa_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de reservas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReservaCollection"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Crea una nueva reserva",
                "operationId": "982e9803d84704d154c36082a018d4e5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreReservaRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reserva creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReservaResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/reservas/{reserva}": {
            "get": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Muestra una reserva específica",
                "operationId": "2398ade4d4df3f92274444d6cb473653",
                "parameters": [
                    {
                        "name": "reserva",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reserva encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReservaResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Reserva no encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Actualiza una reserva existente",
                "operationId": "6088b63d284c0a7f1c9b90afc8bd5b87",
                "parameters": [
                    {
                        "name": "reserva",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateReservaRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reserva actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReservaResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Reserva no encontrada"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Elimina una reserva",
                "operationId": "6228659039fb00362b1498c531173665",
                "parameters": [
                    {
                        "name": "reserva",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Reserva eliminada"
                    },
                    "404": {
                        "description": "Reserva no encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Mesa": {
                "required": [
                    "numero",
                    "capacidad",
                    "estado"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "numero": {
                        "type": "integer",
                        "example": 1
                    },
                    "capacidad": {
                        "type": "integer",
                        "example": 4
                    },
                    "estado": {
                        "type": "string",
                        "enum": [
                            "disponible",
                            "ocupada",
                            "reservada"
                        ],
                        "example": "disponible"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Reserva": {
                "required": [
                    "mesa_id",
                    "fecha",
                    "hora",
                    "cliente_nombre",
                    "cliente_email",
                    "numero_personas"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "mesa_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "fecha": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-12-25"
                    },
                    "hora": {
                        "type": "string",
                        "format": "time",
                        "example": "20:00:00"
                    },
                    "cliente_nombre": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "cliente_email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "cliente_telefono": {
                        "type": "string",
                        "example": "+1234567890"
                    },
                    "numero_personas": {
                        "type": "integer",
                        "example": 4
                    },
                    "estado": {
                        "type": "string",
                        "enum": [
                            "pendiente",
                            "confirmada",
                            "cancelada",
                            "completada"
                        ],
                        "example": "pendiente"
                    },
                    "notas": {
                        "type": "string",
                        "example": "Celebración de cumpleaños"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "mesa": {
                        "$ref": "#/components/schemas/Mesa"
                    }
                },
                "type": "object"
            },
            "ReservaResource": {
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/Reserva"
                    }
                },
                "type": "object"
            },
            "StoreReservaRequest": {
                "required": [
                    "mesa_id",
                    "fecha",
                    "hora",
                    "cliente_nombre",
                    "cliente_email",
                    "numero_personas"
                ],
                "properties": {
                    "mesa_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "fecha": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-12-25"
                    },
                    "hora": {
                        "type": "string",
                        "format": "time",
                        "example": "20:00:00"
                    },
                    "cliente_nombre": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "cliente_email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "cliente_telefono": {
                        "type": "string",
                        "example": "+1234567890"
                    },
                    "numero_personas": {
                        "type": "integer",
                        "example": 4
                    },
                    "notas": {
                        "type": "string",
                        "example": "Celebración de cumpleaños"
                    }
                },
                "type": "object"
            },
            "UpdateReservaRequest": {
                "properties": {
                    "mesa_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "fecha": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-12-25"
                    },
                    "hora": {
                        "type": "string",
                        "format": "time",
                        "example": "20:00:00"
                    },
                    "cliente_nombre": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "cliente_email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "cliente_telefono": {
                        "type": "string",
                        "example": "+1234567890"
                    },
                    "numero_personas": {
                        "type": "integer",
                        "example": 4
                    },
                    "estado": {
                        "type": "string",
                        "enum": [
                            "pendiente",
                            "confirmada",
                            "cancelada",
                            "completada"
                        ],
                        "example": "confirmada"
                    },
                    "notas": {
                        "type": "string",
                        "example": "Celebración de cumpleaños"
                    }
                },
                "type": "object"
            },
            "ReservaCollection": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Reserva"
                        }
                    },
                    "links": {
                        "properties": {
                            "first": {
                                "type": "string"
                            },
                            "last": {
                                "type": "string"
                            },
                            "prev": {
                                "type": "string",
                                "nullable": true
                            },
                            "next": {
                                "type": "string",
                                "nullable": true
                            }
                        },
                        "type": "object"
                    },
                    "meta": {
                        "properties": {
                            "current_page": {
                                "type": "integer"
                            },
                            "from": {
                                "type": "integer"
                            },
                            "last_page": {
                                "type": "integer"
                            },
                            "path": {
                                "type": "string"
                            },
                            "per_page": {
                                "type": "integer"
                            },
                            "to": {
                                "type": "integer"
                            },
                            "total": {
                                "type": "integer"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}